Thu, May 11, 2023, 4:16 PM -07:00

+++ week-1/pset/credit/test.c
@@ -18,7 +18,7 @@ int main(void)
     for (int i = 0; i < len; i++)
     {
         digit[i] = 0;
-        digit[i] = cc_num[i];
+        digit[i] = (int) cc_num[i];
         printf("Digit: %i\n", digit[i]);
 
     }
